requirements.txt
opencv-python
google-generativeai
numpy
gradio
python-dotenv
requests
Pillow
paho-mqtt
pydantic